removed all code
again...the second time.  I am not giving up though, but I think there is just a bit too much snow on the mountains right now.  Need to look at it with the intention of finishing it within one day.  Should not be an issue when my back does not feel like it has 44 hernias.