abortion of concept
I was going to create an entire class dedicated to managing tables and columns etc, but I find it useless right now, for the sake of interest am I going to focus on the Help Manager GUI for now.  There should be many classes I can put together along the way.