replacement
I am currently working on a design for the third release of this framework. The current one works as it should but it has many components which were in the original framework. Therefore will I be recoding one. It will be fully from scratch again. Soon. For now, this has to do. A lot has changed offline, expect some cool updates in a few months.